requirements.txt
altair==5.5.0
asttokens==3.0.0
attrs==24.2.0
Authlib==1.6.0
awscrt==0.24.1
AWSIoTPythonSDK==1.5.4
awsiotsdk==1.22.2
black==25.1.0
blinker==1.9.0
boto3==1.38.22
botocore==1.38.22
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.2
cycler==0.12.1
DateTime==5.5
debugpy==1.8.14
decorator==5.2.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.16.1
fiona==1.10.1
flake8==7.1.2
Flask==3.1.1
fonttools==4.57.0
fsspec==2025.3.2
GeoAlchemy2==0.17.1
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
narwhals==1.14.3
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.1.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.7
plotly==5.24.1
pluggy==1.6.0
prompt_toolkit==3.0.51
protobuf==5.29.0
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pycodestyle==2.12.1
pycparser==2.22
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.18.0
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
pyright==1.1.394
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
s3transfer==0.13.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==79.0.1
shapely==2.1.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.36
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.40.2
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trange==0.1.1
triton==3.3.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.141
ultralytics-thop==2.0.14
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
xlrd==2.0.1
zope.interface==7.2
